+++ src/background/contentScript.ts
@@ -19,12 +19,14 @@ import pDefer from "p-defer";
 import { injectContentScript } from "webext-content-scripts";
 import { getAdditionalPermissions } from "webext-additional-permissions";
 import { patternToRegex } from "webext-patterns";
-import { ENSURE_CONTENT_SCRIPT_READY } from "@/messaging/constants";
-import { isRemoteProcedureCallRequest } from "@/messaging/protocol";
+import { isRemoteProcedureCallRequest } from "@/pageScript/messenger/pigeon";
 import { expectContext } from "@/utils/expectContext";
 import pTimeout from "p-timeout";
 import type { Target } from "@/types";
-import { getTargetState } from "@/contentScript/ready";
+import {
+  getTargetState,
+  ENSURE_CONTENT_SCRIPT_READY,
+} from "@/contentScript/ready";
 import { memoizeUntilSettled } from "@/utils";
 
 /** Checks whether a URL will have the content scripts automatically injected */
