+++ src/blocks/readers/window.ts
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { withReadWindow } from "@/messaging/protocol";
+import { withReadWindow } from "@/pageScript/messenger/api";
 import { type ReaderOutput } from "@/core";
 import { registerFactory } from "@/blocks/readers/factory";
