+++ src/pageEditor/sidebar/SidebarExpanded.tsx
@@ -37,9 +37,9 @@ import {
 } from "@/pageEditor/slices/editorSelectors";
 import { useDispatch, useSelector } from "react-redux";
 import {
-  getRecipeById,
   getIdForElement,
   getRecipeIdForElement,
+  getRecipeById,
 } from "@/pageEditor/utils";
 import useSaveRecipe from "@/pageEditor/hooks/useSaveRecipe";
 import useResetRecipe from "@/pageEditor/hooks/useResetRecipe";
