+++ src/pageEditor/tabs/effect/useDocumentPreviewRunBlock.ts
@@ -26,7 +26,7 @@ import {
   selectParentBlockInfo,
 } from "@/pageEditor/slices/editorSelectors";
 import { getErrorMessage, type SimpleErrorObject } from "@/errors/errorHelpers";
-import { type SerializableResponse } from "@/messaging/protocol";
+import { type SerializableResponse } from "@/pageScript/messenger/pigeon";
 import { createSlice, type PayloadAction } from "@reduxjs/toolkit";
 import { useDebouncedCallback } from "use-debounce";
 import { runRendererBlock } from "@/contentScript/messenger/api";
