+++ src/pageScript/frameworks/adapters.ts
@@ -19,7 +19,7 @@ import emberAdapter from "./contrib/ember";
 import angularjsAdapter from "@/pageScript/frameworks/contrib/angularjs";
 import reactAdapter from "./contrib/react";
 import vueAdapter from "./contrib/vue";
-import { type Framework } from "@/messaging/constants";
+import { type Framework } from "@/pageScript/messenger/constants";
 import {
   type ReadableComponentAdapter,
   type WriteableComponentAdapter,
